requirements.txt
# VERSION 2

flask
flask_sqlalchemy
flask_login
flask_wtf
email_validator
python-dotenv
flask_mail
psycopg2-binary
stripe
reportlab
qrcode
pillow
wtforms-sqlalchemy

# VERSION 2